requirements.txt
falcon==2.0.0
falcon-cors==1.1.7
tensorflow==2.3.1
opencv-python==4.1.2.30
gunicorn==20.0.4
numpy==1.16.4
pillow==7.0.0
pandas==1.1.0
scikit-learn==0.23.2
h5py==2.10.0
